+++ webpack.config.babel.js
@@ -35,7 +35,10 @@ const plugins = [
     template: './src/index.html'
   }),
   new CopyPlugin({
-    patterns: [{ from: 'src/stylesheets', to: 'stylesheets' }]
+    patterns: [
+      { from: 'src/stylesheets', to: 'stylesheets' },
+      { from: 'src/static', to: 'static' }
+    ]
   })
 ];
 
@@ -52,6 +55,10 @@ export default {
       {
         directory: resolve(__dirname, 'src/stylesheets'),
         publicPath: '/stylesheets'
+      },
+      {
+        directory: resolve(__dirname, 'src/static'),
+        publicPath: '/static'
       }
     ],
     compress: dev,